Create The animal morphing siblings episode 1
A script for a pilot episode for a tv series about a 14 year old boy named Liam and his 12 year old sister Abby who can morph into animals. Liam can turn into a wolf and Abby into a lioness. The episode begins with Liam and Abby’s parents sending the siblings to get rakes out of the shed for them. In the shed Abby knocks over a jar of triple A sending her and Liam to sleep. When they don’t return their parents go to the shed and see a wolf and a lioness. They realise that the wolf is Liam and the lioness is Abby. Liam and Abby hunt a rabbit and get a drink from the hose pipe before settling down to sleep. When they wake up they find they are humans again. Later that day the family is visited by David a teacher at Westside College a secret boarding school for superhumans. David explains that superhumans are kids with special powers that are believed to be the next stage of human evolution. David explains that at Westside Liam and Abby will be housed with the 198 other superhumans and will learn how to control their powers as well as continuing their normal studies. David also says they and their powers will be studied by scientists.